backend/requirements-dev.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7  # Para documentación de la API
inflection==0.5.1  # Puede ser útil para transformar nombres en Django
packaging==24.1  # Útil para manejar paquetes
pillow==10.4.0
psycopg2-binary==2.9.9  # Puede ser útil en desarrollo
PyJWT==2.9.0  # Si necesitas manejar JWT
python-decouple==3.8
pytz==2024.2  # Para manejo de zonas horarias
PyYAML==6.0.2  # Si usas YAML en tu proyecto
sqlparse==0.5.0  # Para manejo de SQL
typing_extensions==4.12.2  # Útil si usas anotaciones de tipo
uritemplate==4.1.1  # Si necesitas trabajar con plantillas de URI
whitenoise==6.7.0
